+++ src/main/java/com/viglet/shiohara/ecommerce/ShEcomProductBean.java
@@ -1,0 +1,36 @@
+package com.viglet.shiohara.ecommerce;
+
+public class ShEcomProductBean {
+
+	private String name;
+	
+	private String description;
+	
+	private double value;
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+	public double getValue() {
+		return value;
+	}
+
+	public void setValue(double value) {
+		this.value = value;
+	}
+	
+	
+}