+++ src/main/java/com/viglet/shiohara/mongodb/ShMongoConfig.java
@@ -1,0 +1,5 @@
+package com.viglet.shiohara.mongodb;
+
+public class ShMongoConfig {
+
+}