+++ src/main/resources/js/client-side/src/feature/post/ShPostEditCtrl.js
@@ -29,7 +29,16 @@ shioharaApp.controller('ShPostEditCtrl', [
 			$scope.shPost = shPostResource.get({
 				id : $scope.postId
 			}, function() {
+				angular
+				.forEach($scope.shPost.shPostAttrs,
+						function(shPostAttr, key) {
+							if (angular.equals(shPostAttr.shPostTypeAttr.shWidget.type, "JSON")) {
+								shPostAttr.strValue =  JSON.parse(shPostAttr.strValue);
+							}
+				});
+				
 				if ( $scope.shPost.shFolder != null) {
+					
 					$scope.folderId = $scope.shPost.shFolder.id;
 				$scope
 				.$evalAsync($http
@@ -75,6 +84,11 @@ shioharaApp.controller('ShPostEditCtrl', [
 				angular
 						.forEach($scope.shPost.shPostAttrs,
 								function(shPostAttr, key) {
+							//console.log("Type: " + shPostAttr.shPostTypeAttr.shWidget.type);
+									if (angular.equals(shPostAttr.shPostTypeAttr.shWidget.type, "JSON")) {
+										shPostAttr.strValue =  JSON.stringify(shPostAttr.strValue);
+										console.log(shPostAttr.strValue);
+									}
 									promiseFiles
 											.push(uploadFile(
 													shPostAttr,
@@ -87,6 +101,13 @@ shioharaApp.controller('ShPostEditCtrl', [
 						.then(
 								function(dataThatWasPassed) {
 				$scope.shPost.$update(function() {
+					angular
+					.forEach($scope.shPost.shPostAttrs,
+					function(shPostAttr, key) {					
+								if (angular.equals(shPostAttr.shPostTypeAttr.shWidget.type, "JSON")) {
+									shPostAttr.strValue =  JSON.parse(shPostAttr.strValue);
+								}
+					});
 					 Notification.warning('The ' + $scope.shPost.title +' Post was updated.');
 				});
 								});