+++ src/main/resources/js/server-side/shObject.js
@@ -27,9 +27,9 @@ viglet.shiohara.shObject = function() {
 	 * @desc Returns Form from PostType
 	 * @public
 	 */
-	this.formComponent = function(shPostTypeName) {
-		return shFormComponent.byPostType(shPostTypeName, request);
-	},
+	this.formComponent = function(shPostTypeName, shObjectId) {
+		return shFormComponent.byPostType(shPostTypeName, shObjectId, request);
+	},	
 	/**
 	 * @desc Returns Search Result
 	 * @public